+++ python/cudf/cudf/core/column/string.py
@@ -797,17 +797,23 @@ class StringColumn(column.TypedColumnBase):
     @property
     def is_monotonic_increasing(self):
         if not hasattr(self, "_is_monotonic_increasing"):
-            self._is_monotonic_increasing = _string_column_binop(
-                self[1:], self[:-1], "ge"
-            ).all()
+            if self.has_null_mask:
+                self._is_monotonic_increasing = False
+            else:
+                self._is_monotonic_increasing = libcudf.issorted.issorted(
+                    columns=[self]
+                )
         return self._is_monotonic_increasing
 
     @property
     def is_monotonic_decreasing(self):
         if not hasattr(self, "_is_monotonic_decreasing"):
-            self._is_monotonic_decreasing = _string_column_binop(
-                self[1:], self[:-1], "le"
-            ).all()
+            if self.has_null_mask:
+                self._is_monotonic_decreasing = False
+            else:
+                self._is_monotonic_decreasing = libcudf.issorted.issorted(
+                    columns=[self], descending=[1]
+                )
         return self._is_monotonic_decreasing
 
     @property